move all code from parser to lexer
