+++ Makefile
@@ -6,7 +6,7 @@
 #    By: mzhivoto <mzhivoto@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/11/10 14:19:08 by mzhivoto          #+#    #+#              #
-#    Updated: 2025/04/17 17:22:39 by mzhivoto         ###   ########.fr        #
+#    Updated: 2025/04/21 16:42:18 by mzhivoto         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -45,7 +45,7 @@ PARSING_PATH = $(SRCS_PATH)/parser
 # Files
 MAIN = main.c
 BUILTINS = pwd.c
-PARSING = parser.c split.c utils.c
+LEXER = lexer.c split.c utils.c lexer_format.c operator_check.c
 
 
 LIBFT := $(LIBFT_PATH)/libft.a
