+++ src/lexer/utils.c
